+++ pyproject.toml
@@ -56,6 +56,7 @@ pydoc-markdown = "^4.8.2"
 pandas-stubs = "^2.1.1.230928"
 tenacity = "^8.2.3"
 inquirer = "^3.1.3"
+pydub = "^0.25.1"
 
 [tool.poetry.extras]
 trainers = ["peft","trl","bitsandbytes","scipy"]
